requirements.txt
streamlit
pymongo
nltk
PyPDF2
python-docx
google-generativeai
python-dotenv
